Start sketching out coercion (type-eq evidence) production
My initial approach will be to assert that any two GDP proofs are
equivalent, then figure out how to translate GDP logic types into GHC
coercions (yes, this breaks consistency and will allow me to prove
contradictions).

Added the debug helper `pprCt` to render the different constraint types
in a way that tells me which constructor they used. This showed me that
my `And This That ~# That` was an "irreducible constraint", so I added a
little more info to that branch.