Implement super basic solver (green tests!!)
Broke out the logic into a new module and wrote a basic `solve` function
that can pick apart an `And` and recognize if the `rhs` of the goal is
one of the members of the conjunction. No recursion, no other logical
operators besides `And`.

BUT it's enough for the tests to pass!